% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superscript.R
\name{addSuperScriptToCell}
\alias{addSuperScriptToCell}
\title{Add superscipts/subscripts to Excel files}
\usage{
addSuperScriptToCell(
  wb,
  sheet,
  row,
  col,
  text,
  superScriptText,
  position = nchar(text),
  superOrSub = TRUE,
  size = "10",
  colour = "000000",
  font = "Arial",
  family = "2",
  bold = FALSE,
  italic = FALSE,
  underlined = FALSE
)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A sheet within the workbook object}

\item{row}{The row number you want to write to}

\item{col}{The column number you want to write to}

\item{text}{The text (not including the superscript) you want written.}

\item{superScriptText}{The text you want in the superscript}

\item{position}{A number specifying how far along in the text you want the superscript to occur. Defaults to nchar(text), (ie, the last position)}

\item{superOrSub}{TRUE or FALSE is you want to return a superscript or a subscript. Defaults to superscript (TRUE)}

\item{size}{The size of the font. Defaults to 10}

\item{colour}{The hex code of the colour of the text. Defaults to black (000000)}

\item{font}{The font. Defaults to Arial}

\item{family}{Not sure what this is for. Leave it alone at 2 if you don't know either}

\item{bold}{TRUE or FALSE if you want text to be bold}

\item{italic}{TRUE or FALSE if you want text to be italic}

\item{underlined}{TRUE or FALSE if you want text to be underlined}
}
\description{
Writes text to a cell within a wookbook object and adds a superscript or a subscript.
}
\examples{
\dontrun{
addSuperScriptToCell(wb = wb,
 sheet = "2.1",
 row = 1,
 col =1,
 text = "Table 2.1: First prison receptions by type of first reception, sentence length and sex",
 superScriptText = "(1)",
 position = 34,
 superOrSub = TRUE,
 size = '12',
 colour = '000000',
 font = 'Arial',
 family = '2',
 bold = TRUE,
 italic = FALSE,
 underlined = FALSE)
}

}
