% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_image.R
\name{arrow_image}
\alias{arrow_image}
\title{Chooses arrow image for table on bulletin front page}
\usage{
arrow_image(x)
}
\arguments{
\item{x}{Direction of arrow (up, down, nochange)}
}
\description{
Function to format the correct path to arrow inages for bulletin front page. For this you need to have your .png files stored in a folder called 'images' in the main repo directory. See CCSQ_RAP for an example
}
\examples{
arrow_image("up")

}
