% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_utf.R
\name{arrow_utf}
\alias{arrow_utf}
\title{Output UTF Arrows}
\usage{
arrow_utf(x)
}
\arguments{
\item{x}{Number (usually the difference between two figures)}
}
\value{
UTF arrow depending on whether number is positive, negative or equal to zero
}
\description{
Function to output  Unicode Transformation Format (UTF) arrows depending on the number provided.
}
\examples{
arrow_utf(5)

}
