% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_number.R
\name{as_number}
\alias{as_number}
\title{Takes a string and removes all comma delimeters and '\%' signs, before converting to a numerical format.}
\usage{
as_number(x)
}
\arguments{
\item{x}{A formatted number string}
}
\description{
Takes a string and removes all comma delimeters and '\%' signs, before converting to a numerical format.
}
\examples{
as_number("1,000")  #  returns 1000
as_number("10\%")  #  returns 10
}
