% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculations}
\alias{calculations}
\title{Calculate base/core figures for a RAP system}
\usage{
calculations(table, current_pub_year, current_pub_quarter)
}
\arguments{
\item{table}{The table to extract information from. This is assumed to be in a dataframe and tidy format.}

\item{current_pub_year}{This is the current publication year, usually defined in the begining of your RAP system and need to be passed in the function as well.}

\item{current_pub_quarter}{The curent publication quarter, 
this also is usually defined at the begining of the RAP system}
}
\value{
The value returned is a list containing figures per column (of the main dataframe used in the "tables" argument)
and per category. More specifically:

tot_cases_current --> contains the current quarter values for each of the columns in the main data frame  

tot_cases_previous --> contains the previous quarter values for each of the columns in the main data frame

tot_cases_change   --> contains the change between the current quarter values and the previous one

tot_cases_ratio   --> contains the ratio of the cahnge current and previous quarter divided by the value  of the previous quarter

tot_cases_perc_change   --> contains the percentage cahnge between the current and previous quarter values
}
\description{
This function automates to a large extent the common calculations needed for a RAP system. 
For example the current and pevious value for a table column/entry as well as rations and percentage differences.
Curently 5 categories are included - current values, previous values, numeric change between previous and current quarter values,
the ration of the previous and current value and the percentage difference.
}
\examples{

}
\author{
George Papadopoulos
}
