% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_desc.R
\name{change_desc}
\alias{change_desc}
\title{Describe changes over a given time period}
\usage{
change_desc(df, col_name, col_vals_1, col_val_2 = NULL)
}
\arguments{
\item{df}{A dataframe.}

\item{col_name}{The name of the variable to describe the change for as a string.}

\item{col_vals_1}{A vector specifying the name of the column containing time periods (as a string) and the two corresponding values you want to describe a change between.}

\item{col_val_2}{A vector containing an additional column name (as a string) and value used to filter table rows on, if required.}
}
\value{
A character object.
}
\description{
Returns a character string describing whether a variable has increased or decreased over a given period of time.
}
\examples{
df1 <- data.frame("yrqtr" = c("2015 Q4", "2015 Q4", "2015 Q4", "2016 Q1", "2016 Q1", "2016 Q1"),
"type" = c("receipts", "disposals", "outstanding", "receipts", "disposals", "outstanding"),
"total" = c(3,2,10,1,5,10)
)

df2 <- data.frame("yrqtr" = c("2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = c(10,4,5,10)
)

change_desc(df2, "receipts", c("yrqtr", "2020 Q2", "2020 Q3")) # returns "increased"
change_desc(df2, "receipts", c("yrqtr", "2020 Q1", "2020 Q2")) # returns "decreased"
change_desc(df2, "receipts", c("yrqtr", "2020 Q1", "2020 Q4")) # returns "did not change"
change_desc(df1, "total", c("yrqtr", "2015 Q4", "2016 Q1"), c("type", "disposals"))
# returns "increased"
change_desc(df1, "total", c("yrqtr", "2015 Q4", "2016 Q1"), c("type", "receipts"))
# returns "decreased"
change_desc(df1, "total", c("yrqtr", "2015 Q4", "2016 Q1"), c("type", "outstanding"))
# returns "did not change"

}
