% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_period_formatting.R
\name{check_period_formatting}
\alias{check_period_formatting}
\title{Checks period is always in "Q4 2016" format}
\usage{
check_period_formatting(period)
}
\arguments{
\item{period}{Time period. More specifically, quarter and year}
}
\description{
Checks period is always in "Q4 2016" format
}
\examples{
check_period_formatting("2016 Q3") # returns 0 i.e. does not pass the test
check_period_formattion("Q3 2016") # returns 1 i.e. passes test
}
