% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_value.R
\name{choose_value}
\alias{choose_value}
\title{Choose a value from a dataframe}
\usage{
choose_value(df, var_name, additional_name, additional_value)
}
\arguments{
\item{df}{A dataframe.}

\item{var_name}{Name of the variable we want to extract a value from, as a string.}

\item{additional_name}{The name of a column used to filter the dataframe on, as a string.}

\item{additional_value}{A value in the column to filter the dataframe for.}
}
\description{
Returns the value of a variable in a dataframe, for a given value of another variable (or 2 variables) in the dataframe.
}
\examples{

df <- data.frame("yrqtr" = c("2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = seq(10, 40, 10)
)

choose_value(df, "receipts", "yrqtr", "2020 Q1") # returns 10

}
