% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_series.R
\name{col_series}
\alias{col_series}
\title{Choosing a colour and its number of shades}
\usage{
col_series(colour, class)
}
\arguments{
\item{colour}{Choices of 6 different colours: blue, green, grey, orange, purple and red.}

\item{class}{Number of data classes (different shades) of the chosen colour. The number of data classes can only be between 3 and 9.}
}
\value{
Different shades of chosen colour
}
\description{
You can choose between 6 different colours and how many different shades.
}
\examples{

}
