% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_text.R
\name{composite_text}
\alias{composite_text}
\title{Helper function to assist in the creating multiple sentences}
\usage{
composite_text(comp_data)
}
\arguments{
\item{comp_data}{A data frame comprised of 3 columns,\code{categories}, \code{change} and \code{perc}. These refer to the following:

\itemize{

\item \code{categories}{This column details the names of the categories to include in the multistatement}

\item \code{change} {The corresponding change for each of the categories}

\item \code{perc} {The percentage associated with the hange in the previous column }

}}
}
\value{
A text variable that is to be used to contruct the multistatement structure.
This is stylised for use in later stages of the process
}
\description{
This function is used as a helper mechanism to allow for simpler implementation of the
multistatement function. The main result of the \code{composite_text} function is to create a text component/variable
for use in the multistatement parent funciton. It will align the \code{categories} \code{change} and \code{perc} so that each will in order
}
\examples{

df = tribble(
         ~categories, ~change, ~perc,
         "a",   -3, "10\%",
         "b",   2, "12",
         "c",   1, "5\%",
           )

composite_text(df)

}
