% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_type.R
\name{date_type}
\alias{date_type}
\title{This function takes standard number format and returns a number format of the users choosing}
\usage{
date_type(date, type)
}
\arguments{
\item{date}{standard number format}

\item{type}{specifies the format that the date should be returned in. Standared R date types.}
}
\description{
This function takes standard number format and returns a number format of the users choosing
}
\examples{
date_type("2012-12-31", "\%Y")  # returns 2012
}
