% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2numeric.R
\name{format2numeric}
\alias{format2numeric}
\title{Converting a character value to numeric for use in a RAP system}
\usage{
format2numeric(char_value, separator = ",")
}
\arguments{
\item{separator}{The character to separate the digits. Default here is the comma ","}

\item{value}{The value to be converted}
}
\value{
The value returned is a numeric ready to be used by the RAP system in place.
}
\description{
This is the complement function to `format_num()` and it will transform a corresponding character value back to numeric
}
\examples{

format2numeric(format_num(100222))

format2numeric(format_num(200344445.3322, accuracy = 0.01))

}
\author{
George Papadopoulos
}
