% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Formats numbers with a comma delimiter and values of less than 10 are represented as words}
\usage{
format_num(number)
}
\arguments{
\item{number}{A number}
}
\description{
Formats numbers with a comma delimiter and values of less than 10 are represented as words
}
\examples{
format_num("1000") # returns "1,000"
format_num("1") # returns 'one'
}
