% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_perc.R
\name{format_perc}
\alias{format_perc}
\title{Format percentages}
\usage{
format_perc(fraction)
}
\arguments{
\item{fraction}{A fraction}
}
\value{
A character object.
}
\description{
Convert fractions to percentages.
}
\details{
Generic method to convert tables into wide format
}
\examples{

format_perc(0.1) #returns 10\%

}
