% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incdec.R
\name{incdec}
\alias{incdec}
\title{Returns a character string that quantifies a change.}
\usage{
incdec(value, tense)
}
\arguments{
\item{value}{value to be quantified}

\item{tense}{which tense you want the quantifying string to be in. Either
"past", "present", "singular" or "plural".}
}
\description{
Returns a character string that quantifies a change.
}
\examples{
incdec("-1", "past")  # returns "decreased by"

}
