% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indef_article.R
\name{indef_article}
\alias{indef_article}
\title{A Function to assign the indefinite article before a word}
\usage{
indef_article(text_var, caps = F)
}
\arguments{
\item{text_var}{The word before which the article will be placed}
}
\value{
A text variable conaining the multistaement ready to be used in the RAP system
}
\description{
This function attempts to ssign the indefinite article before a given text string.
Currently this will work with words but not numbers or anythig that cannot be used by regex
which is the core ext recognition engine used.

#'
}
\examples{


article_indef("word", caps=F)

article_indef("A", caps=F)


}
