% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_quarter.R
\name{latest_quarter}
\alias{latest_quarter}
\title{Returns name of latest quarter, for within the bulletin.}
\usage{
latest_quarter()
}
\description{
Returns name of latest quarter, for within the bulletin.
}
