% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{moj_colours}
\alias{moj_colours}
\title{MoJ colours}
\usage{
moj_colours(...)
}
\arguments{
\item{...}{Colour names as character strings.}
}
\value{
Returns a named character vector.
}
\description{
Provides names and hex codes for MoJ corporate colours. Calling the function
with no arguments returns the name and hex code of every colour. If the names
of specific colours are passed as arguments, the function returns the names
and hex codes for those colours alone. This function originates from the
mojchart package, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\examples{
# All colours
moj_colours()

# Selected colours
moj_colours("mojblue", "mojgreen")
}
\seealso{
MoJ branding intranet page:
  \url{https://intranet.justice.gov.uk/guidance/communications/branding-templates/}

Other mojchart functions: 
\code{\link{mojchart_palette}()},
\code{\link{palette_colours}()},
\code{\link{scales}},
\code{\link{scheme_names}()},
\code{\link{theme_gss}()}
}
\concept{available colours}
\concept{mojchart functions}
