% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mojchart_palette}
\alias{mojchart_palette}
\title{Single colour palette}
\usage{
mojchart_palette(n, scheme)
}
\arguments{
\item{n}{The required number of colours in the palette, from one to six.}

\item{scheme}{The name of a colour scheme as a character string. Run
`scheme_names()` for the available options.}
}
\description{
Returns a single colour palette as a named character vector. This function
originates from the mojchart package, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\examples{
mojchart_palette(5, scheme = "vibrant1")
}
\seealso{
Other mojchart functions: 
\code{\link{moj_colours}()},
\code{\link{palette_colours}()},
\code{\link{scales}},
\code{\link{scheme_names}()},
\code{\link{theme_gss}()}
}
\concept{mojchart functions}
