% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojquarter.R
\name{mojquarter}
\alias{mojquarter}
\title{Convert dates to publication quarters}
\usage{
mojquarter(date, type)
}
\arguments{
\item{date}{Date in the form of 'YYYY-MM-DD'}

\item{type}{Input 1 for calendar quarter or 4 for financial quarter}
}
\value{
Either a calendar or financial quarter
}
\description{
Function converting dates to publication quarters
}
\examples{
mojquarter('2020-03-31', 4)

}
