% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tables.R
\name{multiple_cell_merge}
\alias{multiple_cell_merge}
\title{Merges multiple columns of cells in an Excel sheet, by row}
\usage{
multiple_cell_merge(wb, sheet, rows, cols)
}
\arguments{
\item{wb}{openxlsx workbook object}

\item{sheet}{string. Sheet name in the wb object}

\item{rows}{numeric vector. Row numbers in the wb object that you want to apply the merge formatting to}

\item{cols}{numeric vector. Column numbers in the wb object that you want to merge together.}
}
\description{
Merges multiple columns of cells in an Excel sheet, by row
}
