% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{palette_colours}
\alias{palette_colours}
\title{Palette colours}
\usage{
palette_colours(...)
}
\arguments{
\item{...}{Colour names as character strings.}
}
\value{
Returns a named character vector.
}
\description{
Provides names and hex codes for colours used in  palettes. If colour names
are passed as arguments, returns those colours. If no arguments are provided,
returns every colour. This function originates from the mojchart package, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\examples{
# All colours
palette_colours()

# Selected colours
palette_colours("mojblue", "yellow1")
}
\seealso{
Other mojchart functions: 
\code{\link{moj_colours}()},
\code{\link{mojchart_palette}()},
\code{\link{scales}},
\code{\link{scheme_names}()},
\code{\link{theme_gss}()}
}
\concept{mojchart functions}
