% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_quarter.R
\name{previous_quarter}
\alias{previous_quarter}
\title{Return a previous quarter}
\usage{
previous_quarter(quarter, quarters_ago = 1)
}
\arguments{
\item{quarter}{The latest quarter as a string, starting with the year, a 'q' or 'Q', then the quarter number (1,2,3 or 4).}

\item{quarters_ago}{Number of quarters to go back. Default is 1.}
}
\value{
A character object.
}
\description{
Given the latest quarter as a string with the year first and then the quarter number, e.g. "2020 Q4", returns a
string in the same format representing the previous quarter, or previous quarters going further back if desired.
}
\examples{
previous_quarter("2020 Q1") # returns "2019 Q4"
previous_quarter("2020 Q1", 3) # returns "2019 Q2"
}
