% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dater.r
\name{pub_date}
\alias{pub_date}
\title{Returns the next publication date}
\usage{
pub_date(pub_day = 0, pub_week = 0, first_pub = 0, frequency = 0,
  bi_start = NA, mago = 0)
}
\arguments{
\item{pub_day}{what day of the week is your publication (3 letter starting with capital - Mon, Tue, Wed etc.)}

\item{pub_week}{which week of the month do you publish (1, 2, 3, 4 or 5 for the last week)}

\item{first_pub}{in what month is the first publication of the calendar year (if quarterly this should be 1, 2 or 3)}

\item{frequency}{how often do you publish (monthly, quarterly, annually, biennially)}

\item{bi_start}{biennial publications only is the year of publication "odd" or "even"}

\item{mago}{Reduce the date by this many months}
}
\description{
Takes the publication day, week and other parameters and returns the next publication date.
Can be used to find previous publications by specifying the months ago parameter.
}
\details{
This will take today's date and find the next publication date for a publication specified using the below parameters.
It does a check to make to make sure that it is not a past date unless backdating is specified.
}
\examples{
pub_date(pub_day = "Thu", pub_week = 2, first_pub = 2, frequency = "quarterly", bi_start = NA , mago = 0)

}
