% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarters_ago.R
\name{quarters_ago}
\alias{quarters_ago}
\title{Returns period going back a specified number of quarters}
\usage{
quarters_ago(quarters_ago)
}
\arguments{
\item{quarters_ago}{Number of quarters to go back}
}
\description{
Returns period going back a specified number of quarters
}
\examples{
params$year <- 2016
params$quarter <- Q4

quarters_ago(1) # Returns "Q3 2016" 
quarters_ago(4) # Returns "Q4 2015"
}
