% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cases_to_data.R
\name{read_cases_to_data}
\alias{read_cases_to_data}
\title{Read case-level SAS or CSV file from S3 into an aggregated dataset}
\usage{
read_cases_to_data(format, path, varlist)
}
\arguments{
\item{format}{Format of the dataset in S3 (either "SAS" or "CSV"). NB Case insensitive input.}

\item{path}{Path to file in S3.}

\item{varlist}{A character vector of variables by which to aggregate the data.}
}
\description{
Reads in a case-level SAS or CSV file stored in S3 and aggregates it by a vector of specified variables.
Adds a variable called 'count' to the dataset to indicate the number of cases in each variable group.
}
\examples{
\dontrun{
read_cases_to_data("SAS","alpha-test-data/popdata.sas7bdat",c("sex","age","region"))
# returns a tibble with variables sex, age, region and count.
}
}
