% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tables.R
\name{save_formatted_table}
\alias{save_formatted_table}
\title{Writes multiple data tables and accompanying notes to an Excel sheet, with formatting.}
\usage{
save_formatted_table(
  workbook,
  sheet_name,
  tables,
  notes,
  starting_row,
  quarterly_format = NULL
)
}
\arguments{
\item{workbook}{openxlsx workbook object}

\item{sheet_name}{string. Sheet name in the wb object}

\item{tables}{a list of data.frame objects to write to the sheet.}

\item{notes}{a character vector of notes to write to the sheet}

\item{starting_row}{numeric. Row number on the sheet you would like to start writing the tables at.}

\item{quarterly_format}{numeric vector. List position of the tables in the tables list to apply quarterly formatting to. Default is NULL.}
}
\description{
Writes multiple data tables and accompanying notes to an Excel sheet, with formatting.
}
