% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scales}
\alias{scales}
\alias{scale_colour_moj}
\alias{scale_fill_moj}
\title{ggplot2 scale functions}
\usage{
scale_colour_moj(n, scheme = "vibrant1", order = NULL, ...)

scale_fill_moj(n, scheme = "muted1", order = NULL, ...)
}
\arguments{
\item{n}{The number of colours required, from one to six.}

\item{scheme}{The name of a colour scheme. Run `scheme_names()` for the
available options. "Muted" schemes are generally recommended for the fill
scale and "vibrant" schemes for the colour scale.}

\item{order}{A numeric vector giving the order in which to apply the colours.
`order` must have length `n`. Numbers can be repeated to apply the same
colour to multiple categories in the data.}

\item{...}{Other arguments passed to scale_discrete_manual().}
}
\value{
Returns a ggplot2 scale function.
}
\description{
These functions add a colour palette to a ggplot2 chart. They originate
from the mojchart package. For an overview and examples, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\examples{
library(dplyr)
library(ggplot2)
txhousing \%>\%
  filter(city \%in\% c("Abilene", "Bay Area", "Collin County")) \%>\%
  ggplot(aes(x = date, y = sales, group = city, colour = city)) +
  geom_line(size = 1.5) +
  scale_colour_moj(3, scheme = "vibrant1")

mpg \%>\%
  ggplot(aes(x = factor(year), y = displ, fill = factor(cyl))) +
  geom_col(position = "dodge") +
  theme_gss() +
  scale_fill_moj(4, scheme = "muted1")
}
\seealso{
Other mojchart functions: 
\code{\link{moj_colours}()},
\code{\link{mojchart_palette}()},
\code{\link{palette_colours}()},
\code{\link{scheme_names}()},
\code{\link{theme_gss}()}
}
\concept{mojchart functions}
