% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scheme_names}
\alias{scheme_names}
\title{Colour scheme names}
\usage{
scheme_names()
}
\description{
Returns the names of the available colour schemes. This function originates from
the mojchart package, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\seealso{
Other mojchart functions: 
\code{\link{moj_colours}()},
\code{\link{mojchart_palette}()},
\code{\link{palette_colours}()},
\code{\link{scales}},
\code{\link{theme_gss}()}
}
\concept{mojchart functions}
