% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gss.R
\name{theme_gss}
\alias{theme_gss}
\title{GSS style theme}
\usage{
theme_gss(
  base_size = 20,
  base_line_size = 0.5,
  flipped = FALSE,
  xticks = FALSE,
  xlabel = FALSE
)
}
\arguments{
\item{base_size}{The base font size for chart text. For publication charts on
gov.uk you probably need around size 20. For details of what this parameter
does, see \url{https://ggplot2-book.org/polishing.html}.}

\item{base_line_size}{The base size for lines controlled by the theme
function (i.e. non-data lines).}

\item{flipped}{A logical value indicating whether the chart has flipped axes.}

\item{xticks}{A logical value indicating whether to add tick marks to the x
axis. Only adds ticks if `flipped = FALSE`.}

\item{xlabel}{A logical value indicating whether to show the x axis label.}
}
\value{
Returns a ggplot2 theme function.
}
\description{
A theme based on Government Statistical Service (GSS) guidance:
\url{https://gss.civilservice.gov.uk/policy-store/introduction-to-data-visualisation/}.\cr
This function originates from the mojchart package. For an overview and
examples of suggested usage, see
\url{https://github.com/moj-analytical-services/mojchart}.
}
\examples{
library(dplyr)
library(ggplot2)
txhousing \%>\%
  filter(city \%in\% c("Abilene", "Bay Area", "Collin County")) \%>\%
  ggplot(aes(x = date, y = sales, group = city, colour = city)) +
  labs(title = "Housing sales in Texas",
       subtitle = "2000 to 2015\n\nNumber of sales",
       caption = "Source: txhousing dataset, ggplot2") +
  expand_limits(y = 2000) +
  geom_line(size = 1.5) +
  theme_gss()
}
\seealso{
Other mojchart functions: 
\code{\link{moj_colours}()},
\code{\link{mojchart_palette}()},
\code{\link{palette_colours}()},
\code{\link{scales}},
\code{\link{scheme_names}()}
}
\concept{mojchart functions}
