% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearago_quarter.R
\name{yearago_quarter}
\alias{yearago_quarter}
\title{Returns name of previous quarter, for within the bulletin.}
\usage{
yearago_quarter()
}
\description{
Returns name of previous quarter, for within the bulletin.
}
