% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_image.R
\name{arrow_image}
\alias{arrow_image}
\title{Chooses arrow image for table on bulletin front page}
\usage{
arrow_image(x, image_path = "images")
}
\arguments{
\item{x}{Direction of arrow (up, down, nochange)}

\item{image_path}{The path to the directory containing the arrow images. Default is a folder called 'images' in the repo root directory.}
}
\value{
Path to image of upwards arrow if x is "up", downwards arrow if x is "down" or a no change arrow if x is "nochange".
}
\description{
Function to format the correct path to arrow images for bulletin front page.
}
\examples{
arrow_image("up")

}
