% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_pdf.R
\name{arrow_pdf}
\alias{arrow_pdf}
\title{Output arrow image for table on PDF bulletin front page}
\usage{
arrow_pdf(x, path = "images")
}
\arguments{
\item{x}{Number (usually the difference between two figures)}

\item{path}{The path to the directory containing the arrow images. Default is a folder called 'images' in the repo root directory.}
}
\value{
Path to image of upwards arrow if x is positive, downwards arrow if x is negative or a no change arrow when x is zero.
}
\description{
Function to format the correct path to arrow images for PDF bulletin front page.
}
\examples{
arrow_pdf(5)

}
