% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Calculate changes over a given time period}
\usage{
change(df, var_name, date_col, base_date, second_date, type = "absolute")
}
\arguments{
\item{df}{A dataframe.}

\item{var_name}{The name of the variable to calculate the change for, as a string.}

\item{date_col}{The name of the column containing time periods, as a string.}

\item{base_date}{The base date to calculate the change between.}

\item{second_date}{The second date to calculate the change between.}

\item{type}{Specify either an "absolute" or a "percentage" change. Defaults to "absolute".}
}
\value{
A numeric value.
}
\description{
Calculates either the absolute or percentage change in a given variable over a certain period of time.
}
\examples{

df <- data.frame("yrqtr" = c("2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = seq(10, 40, 10)
)

change(df, "receipts", "yrqtr", "2020 Q2", "2020 Q4") # returns 20
change(df, "receipts", "yrqtr", "2020 Q2", "2020 Q4") # returns "100\%"
}
