% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Calculates the change in a given variable over a given period}
\usage{
change(df, x, y, z = "absolute")
}
\arguments{
\item{df}{Dataset with column called "variable"}

\item{x}{Chosen value from "variable" column}

\item{y}{A number greater than 0. This is the number of quarters ago,
that you want to calculate change from.}

\item{z}{Absolute or percentage change.}
}
\description{
Calculates the change in a given variable over a given period
}
