% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Calculate changes over a given time period}
\usage{
change(df, col_name, val, ago, type = "absolute")
}
\arguments{
\item{df}{A dataframe.}

\item{col_name}{Variable name from the dataframe to calculate the change for.}

\item{val}{Chosen value of the variable to calculate the change for. NB the first instance of the value in the column is matched.}

\item{ago}{The number of time intervals ago, e.g. quarters or years, to calculate the change from. An integer.}

\item{type}{Specify either an "absolute" or "percentage" change.}
}
\value{
A numeric value.
}
\description{
Calculates either the absolute or percentage change in a given variable over a given period of time.
}
\examples{
df <- data.frame("period" = c("2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = seq(10, 40, 10)
)
change(df, "receipts", 40, 2) # returns 20
change(df, "receipts", 40, 1, type = "percentage") # returns "33\%"

}
