% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_desc.R
\name{change_desc}
\alias{change_desc}
\title{Describe changes over a given time period}
\usage{
change_desc(df, col_name, val, ago)
}
\arguments{
\item{df}{A dataframe.}

\item{col_name}{Variable name from the dataframe to calculate the change for.}

\item{val}{Chosen value of the variable to calculate the change for.}

\item{ago}{The number of quarters ago for quarterly data,
or years ago for yearly data, that you want to calculate the change from.}
}
\value{
A character object.
}
\description{
Returns a character string describing whether a variable has increased or decreased over a given period of time.
}
\examples{
df <- data.frame("period" = c("2019 Q4", "2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = c(100, seq(10, 40, 10))
)
change_desc(df, "receipts", 40, 2) # returns "increased"
change_desc(df, "receipts", 40, 4) # returns "decreased"

}
