% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_desc.R
\name{change_desc}
\alias{change_desc}
\title{Describe changes over a given time period}
\usage{
change_desc(df, var_name, date_col, base_date, second_date)
}
\arguments{
\item{df}{A dataframe.}

\item{var_name}{The name of the variable to calculate the change for, as a string.}

\item{date_col}{The name of the column containing time periods, as a string.}

\item{base_date}{The base date to calculate the change between.}

\item{second_date}{The second date to calculate the change between.}
}
\value{
A character object.
}
\description{
Returns a character string describing whether a variable has increased or decreased over a given period of time.
}
\examples{

df <- data.frame("yrqtr" = c("2020 Q1", "2020 Q2", "2020 Q3", "2020 Q4"),
"receipts" = c(10,4,5,10)
)

change_desc(df, "receipts", "yrqtr", "2020 Q2", "2020 Q3") # returns "increased"
change_desc(df, "receipts", "yrqtr", "2020 Q1", "2020 Q2") # returns "decreased"
change_desc(df, "receipts", "yrqtr", "2020 Q1", "2020 Q4") # returns "did not change"

}
