% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_desc.R
\name{change_desc}
\alias{change_desc}
\title{Returns whether a variable has increased or decreased over a given period.}
\usage{
change_desc(df, x, y)
}
\arguments{
\item{df}{Dataset with column called "variable"}

\item{x}{Chosen value from "variable" column}

\item{y}{A number greater than 0. This is the number of quarters ago,
that you want to calculate change from.}
}
\description{
Returns whether a variable has increased or decreased over a given period.
}
