% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_value.R
\name{choose_value}
\alias{choose_value}
\title{Choose values from a dataframe}
\usage{
choose_value(
  df,
  period_col_name,
  period_val,
  value_col_name,
  additional_col_name = NULL,
  additonal_val = NULL
)
}
\arguments{
\item{df}{A dataframe.}

\item{period_col_name}{Name of the column containing time periods.}

\item{period_val}{Time period to choose a corresponding value for.}

\item{value_col_name}{Name of the column containing the variable we want to extract values from.}

\item{additional_col_name}{An additional variable name to match a value for when selecting a value.}

\item{additonal_val}{Value of the additional variable to be matched when selecting a value.}
}
\value{
A numeric value or character object.
}
\description{
Returns the value of a given variable from a dataframe for a given time period, and optionally for
a given value in an additional column.
}
\examples{
df <- data.frame("yrqtr" = c("2015 Q4", "2015 Q4", "2015 Q4", "2016 Q1", "2016 Q1", "2016 Q1"),
"type" = c("receipts", "disposals", "outstanding", "receipts", "disposals", "outstanding"),
"total" = 1:6
)

choose_value(df, "yrqtr", "2015 Q4", "total", "type", "outstanding") # returns 3

}
