% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_value.R
\name{choose_value}
\alias{choose_value}
\title{Picks out selected values from data frame for use in publicatoin commentary}
\usage{
choose_value(data, Period, Variable)
}
\arguments{
\item{data}{Dataset with columns 'period', 'variable' and 'value'}

\item{Period}{Period to be chosen - in "2016 Q1" or "Q1 2016" format}

\item{Variable}{Variable of data to be chosen}
}
\description{
Picks out selected values from data frame for use in publicatoin commentary
}
\examples{
data <- tibble("period" = c("2015 Q4", "2015 Q4", "2015 Q4", "2016 Q1", "2016 Q1", "2016 Q1"),
"variable" = c("receipts", "disposals", "outstanding", "receipts", "disposals", "outstanding"),
"value" = 1:6
)

choose_value(data, "2015 Q4", "outstanding")

}
