% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_type.R
\name{date_type}
\alias{date_type}
\title{Format dates}
\usage{
date_type(date, type)
}
\arguments{
\item{date}{A date in "\%Y\%m\%d" format.}

\item{type}{Specifies the format that the date should be returned in,
using standard R date types.}
}
\value{
A character object.
}
\description{
Transform dates in "\%Y\%m\%d" format, i.e. with date formats in that order but
with arbitrary or no separators, to a specified date format.
}
\examples{
date_type("2012-12-31", "\%Y")  # returns 2012

}
