% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_expenditure.R
\name{format_expenditure}
\alias{format_expenditure}
\title{Format expenditure}
\usage{
format_expenditure(value, format = "m", currency = "£", dp = 1)
}
\arguments{
\item{value}{An amount of expenditure in}

\item{format}{Whether the expenditure is written in millions (m) or thousands (k). Default: millions}

\item{currency}{Default: £'s}

\item{dp}{Number of decimal places to round expenditure to. Default: 1 decimal place}
}
\value{
A character object.
}
\description{
Convert value to amount for publication. By default, the expenditure format is in £m with 1 decimal place.
}
\details{
Generic method to convert value into a £m, £k, $m etc.
}
\examples{

format_expenditure(36573436) #returns £36.6m
format_expenditure(3562744, currency = "$") #returns $3.6m
format_expenditure(3562744, format = "k") #returns £3562.7k
format_expenditure(4654262.65, currency = "£", dp = 3) #returns £4.654m

}
