% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Formating a numeric value to display a comma at the thousands mark for use in a RAP system}
\usage{
format_num(value, big.mark = ",", decimal.mark = ".", accuracy = 0.1)
}
\arguments{
\item{value}{The value to be converted}

\item{big.mark}{The character to separate the digits. Default here is the comma ","}

\item{decimal.mark}{The character to separate the decimal digits. Default here is the dot "."}

\item{accuracy}{The desired accuracy, the number to round up to.
By default this is set to 0.1 and it will round up to the first decimal point.}
}
\value{
The value returned is a character string ready to be used by the RAP system in place.
}
\description{
This function outputs and formats a numeric value to
appear as a character string with commas separating the thousands mark. The function is a wrapper of the "comma()" function from
the "scales" package
}
\examples{

format_num(100222)

format_num(200344445.3322, accuracy = 0.01)

}
