% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_perc.R
\name{format_perc}
\alias{format_perc}
<<<<<<< HEAD
<<<<<<< HEAD
\title{Formating the percentage value for use in a RAP system}
\usage{
format_perc(previous_value, current_value, round_digits = 0,
  times100 = T)
}
\arguments{
\item{previous_value}{Having quarterly changes in mind (between two consecutive years), this is the previous quarter value.}

\item{current_value}{Having quarterly changes in mind (between two consecutive years), this is the current quarter value.}

\item{round_digits}{The number of digits to round up the result, default value is set to zero (0)}

\item{times100}{Whether the result should be presented as a decimal number or a percentage, default flag set to TRUE,
meaning that the result will be in the percentage format}
}
\value{
The value returned is a character string ready to be used by the RAP system in place.
}
\description{
This function outputs and formats the percentage change between a pair of values.
}
\examples{

format_perc(100,120)

format_perc(100,99.5)

}
\author{
George Papadopoulos
}
=======
=======
>>>>>>> 8e0c8be30642eb2d35eb78e8509a82d45d14570f
\title{Format percentages}
\usage{
format_perc(fraction)
}
\arguments{
\item{fraction}{A fraction}
}
\value{
A character object.
}
\description{
Convert fractions to percentages.
}
\details{
Generic method to convert tables into wide format
}
\examples{

format_perc(0.1) #returns 10\%

}
<<<<<<< HEAD
>>>>>>> 8e0c8be30642eb2d35eb78e8509a82d45d14570f
=======
>>>>>>> 8e0c8be30642eb2d35eb78e8509a82d45d14570f
