% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_perc.R
\name{format_perc}
\alias{format_perc}
\title{Format percentages}
\usage{
format_perc(fraction, error_on_NA = TRUE)
}
\arguments{
\item{fraction}{A fraction}

\item{error_on_NA}{A logical}
}
\value{
A character object.
}
\description{
Convert fractions to percentages.
}
\examples{

format_perc(0.1) # returns "10\%"
format_perc(10) # returns "1000\%"
format_perc(NA, error_on_NA = FALSE) # returns NA

}
