% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moj_col.R
\name{moj_col}
\alias{moj_col}
\title{Choosing a colour, its number of data classes and shade}
\usage{
moj_col(colour, x, class = 9)
}
\arguments{
\item{colour}{Choices of 6 different colours: blue, green, grey, orange, purple and red.}

\item{x}{Shade of colour, in order from brightest to darkest.}

\item{class}{Number of data classes (different shades) of the chosen colour. The default number of classes is 9. The number of data classes can only be between 3 and 9}
}
\value{
Colour(s)
}
\description{
You can choose between 6 different colours and their shades. This function can be useful for adding colour to your charts.
}
\examples{

}
