% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojquarter.R
\name{mojquarter}
\alias{mojquarter}
\title{Convert dates to publication quarters}
\usage{
mojquarter(date, type = "c")
}
\arguments{
\item{date}{A date as a string in the format "YYYY-MM-DD". Only last two digits for the year is also valid.}

\item{type}{Either "c" or "f", to specify a calender quarter or a financial quarter. By default, the function outputs a calendar quarter.}
}
\value{
A character object.
}
\description{
Convert dates as a string in "YYYY-MM-DD" format, with arbitrary separators, to either calendar or
financial quarters. N.B. a financial year in the UK starts in April.
}
\examples{
mojquarter('2020-03-31', "f") # returns "2019q4"

}
