% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojquarter.R
\name{mojquarter}
\alias{mojquarter}
\title{Convert dates to publication quarters}
\usage{
mojquarter(date, type = "c")
}
\arguments{
\item{date}{A date in the form "\%Y\%m\%d".}

\item{type}{Either "c" or "f", to specify a calender quarter or a financial quarter.}
}
\value{
A character object.
}
\description{
Convert dates in "\%Y\%m\%d" format, i.e. with date formats in that order but
with arbitrary or no separators, to either calendar or financial quarters (i.e. where the financial
year starts on the 4th month of the year).
}
\examples{
mojquarter('2020-03-31', "f") # returns "2019q4"

}
