% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_quarter.R
\name{previous_quarter}
\alias{previous_quarter}
\title{Returns name of previous quarter, for within the bulletin.}
\usage{
previous_quarter()
}
\description{
Returns name of previous quarter, for within the bulletin.
}
